+++ handbook/src/pages/index.js
@@ -81,7 +81,7 @@ function Banner() {
             >
               一分钟上手
             </Link>
-            <div className="furion-version">v2.17.10</div>
+            <div className="furion-version">v2.17.11</div>
           </div>
         </div>
         <div className="furion-banner-item">
@@ -164,7 +164,7 @@ function Gitee() {
             className={"furion-log-jiao" + (isDarkTheme ? " dark" : "")}
           ></div>
           <div className="furion-log-number">
-            <div style={{ color: "#1fd898" }}>985,038</div>
+            <div style={{ color: "#1fd898" }}>989,329</div>
             <span className={isDarkTheme ? " dark" : ""}>Downloads</span>
           </div>
         </div>